finalização para entrega
