+++ sql_files/posts.sql
@@ -1,7 +1,7 @@
 SET SQL_SAFE_UPDATES = 0;
-DELETE FROM posts WHERE id = "45d9314b-fa9d-45d2-b50b-780511715440";
+/*DELETE FROM posts WHERE id = "45d9314b-fa9d-45d2-b50b-780511715440";
 TRUNCATE TABLE posts;
-DROP TABLE posts;
+DROP TABLE posts;*/
 
 DESCRIBE posts;
