+++ src/controller/PostsController.ts
@@ -4,9 +4,10 @@ import { Authenticator } from '../services/Authenticator';
 import { IdGenerator } from '../services/IdGenerator';
 import { PostDatabase } from '../data/PostDatabase';
 import { FeedDatabase } from '../data/FeedDatabase';
-import dayjs from 'dayjs';
 import { SearchPostDTO } from '../model/Post';
 import { PostBusiness } from '../business/PostBusiness';
+
+import dayjs from 'dayjs';
 
 export default class PostsController {
   public createPost = async (req: Request, res: Response) => {
@@ -108,5 +109,48 @@ export default class PostsController {
     } catch (error) {
       res.status(400).send(error.message);
     }
+  };
+
+  public like = async (req: Request, res: Response) => {
+    try {
+      const token = req.headers.authorization as string;
+      const postToLike = req.body.postToLike;
+
+      const postBusiness = new PostBusiness();
+      await postBusiness.like(token, postToLike);
+
+      res.status(200).send({
+        message: `Você deu "LIKE" no post de ID:${postToLike}`,
+      });
+    } catch (e) {
+      if (e.sqlMessage.includes('ER_DUP_ENTRY')) {
+        res.status(400).send({
+          message: 'Você já curtiu esse post',
+        });
+      }
+      res.status(400).send({
+        message: e.message || e.sqlMessage,
+      });
+    }
+    await BaseDatabase.destroyConnection();
+  };
+
+  public dislike = async (req: Request, res: Response) => {
+    try {
+      const token = req.headers.authorization as string;
+      const postToDislike = req.body.postToDislike;
+
+      const postBusiness = new PostBusiness();
+      await postBusiness.dislike(token, postToDislike);
+
+      res.status(200).send({
+        message: 'Você "deslaikeou" o post com sucesso',
+      });
+    } catch (e) {
+      res.status(400).send({
+        message: e.message,
+      });
+    }
+    await BaseDatabase.destroyConnection();
   };
 }
